Initial Project Upload
The WebLogImport project is a C# Windows app that imports raw IIS
webserver logs for preview and saving to a SQL Server database. To use
the program, a SQL Server table must be created with fields for each of
the 22 fields in the IIS log file along with a 23 field for a domain
name or other identifier to go with the website. This allows for logs
for multiple sites to be retained.